interpolation: split into several files
