+++ opendrift/readers/interpolation/__init__.py
@@ -1,0 +1,3 @@
+from .interpolators import *
+from .structured import ReaderBlock
+
